backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pillow
opencv-python
numpy
pydantic
easyocr
motor
pymongo
PyJWT
bcrypt
